+++ CHANGELOG.md
@@ -2,17 +2,24 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `InjectableServiceKey`. [#4](https://github.com/go-nacelle/service/pull/4)
+- Exported top-level `Inject` method. [#9](https://github.com/go-nacelle/service/pull/9)
+- Added `WithValues` to `Container`. [#9](https://github.com/go-nacelle/service/pull/9)
+
 ### Changed
 
 - Change type of service keys from `string` to `interface{}`. [#4](https://github.com/go-nacelle/service/pull/4)
 - Replaced the `ServiceContainer` interface with `Container`, a struct with the same name and set of methods. [#7](https://github.com/go-nacelle/service/pull/7)
 - Renamed `NewServiceContainer` to `New`. [#7](https://github.com/go-nacelle/service/pull/7)
-- Renamed `Overlay` to `NewOverlay`. [#7](https://github.com/go-nacelle/service/pull/7)
+- Removed `Inject` method from `Container`. [#9](https://github.com/go-nacelle/service/pull/9)
 
 ### Removed
 
-- Removed `MustGet` and `MustSet` methods from the `Container` interface. [#3](https://github.com/go-nacelle/service/pull/3)
+- Removed `MustGet` and `MustSet` methods. [#3](https://github.com/go-nacelle/service/pull/3)
 - Removed mocks package. [#6](https://github.com/go-nacelle/service/pull/6)
+- Removed `Overlay`. [#9](https://github.com/go-nacelle/service/pull/9)
 
 ## [v1.0.2] - 2020-09-30